% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drought.R
\name{drought_county}
\alias{drought_county}
\title{Get county-level drought severity statistics (percent area)}
\usage{
drought_county(state, start = "1/1/2024", end = "12/31/2024")
}
\arguments{
\item{state}{Two-letter state abbreviation (e.g., "CA", "TX") or FIPS code}

\item{start}{Start date as "M/D/YYYY" (e.g., "1/1/2024")}

\item{end}{End date as "M/D/YYYY" (e.g., "12/31/2024")}
}
\value{
tibble: date, county, state, none, d0, d1, d2, d3, d4
}
\description{
Returns weekly drought severity percentages for counties in a given state.
}
